requirements.txt
# Backend Veraleza (FastAPI + Supabase + Pandas)
typing_extensions>=4.6.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
PyJWT>=2.8.0